locust/requirements.txt
locust 
requests 
python-dotenv